requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.8
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
Pillow==9.2.0
pycodestyle==2.9.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.4
